Reducing redundancy in keygen reshare and refresh testing
