+++ src/ecdsa/dkg_ecdsa.rs
@@ -1,12 +1,13 @@
 #[cfg(test)]
 mod test {
-    use frost_secp256k1::{Secp256K1Sha256, Secp256K1Group, Group};
+    use frost_secp256k1::{Group, Secp256K1Group, Secp256K1Sha256};
     type E = Secp256K1Sha256;
 
-    use crate::ecdsa::test::{assert_public_key_invariant, run_keygen, run_refresh, run_reshare};
     use crate::participants::ParticipantList;
     use crate::protocol::Participant;
-    use crate::test::generate_participants;
+    use crate::test::{
+        assert_public_key_invariant, generate_participants, run_keygen, run_refresh, run_reshare,
+    };
     use std::error::Error;
 
     #[test]
@@ -18,7 +19,7 @@ mod test {
         ];
         let threshold = 3;
 
-        let result = run_keygen(&participants, threshold)?;
+        let result = run_keygen::<E>(&participants, threshold)?;
         assert_public_key_invariant(&result)?;
 
         assert!(result.len() == participants.len());
@@ -50,7 +51,7 @@ mod test {
         ];
         let threshold = 3;
 
-        let result0 = run_keygen(&participants, threshold)?;
+        let result0 = run_keygen::<E>(&participants, threshold)?;
         assert_public_key_invariant(&result0)?;
 
         let pub_key = result0[2].1.public_key.to_element();
@@ -78,7 +79,7 @@ mod test {
         let threshold0 = 2;
         let threshold1 = 3;
 
-        let result0 = run_keygen(&participants, threshold0)?;
+        let result0 = run_keygen::<E>(&participants, threshold0)?;
         assert_public_key_invariant(&result0)?;
 
         let pub_key = result0[2].1.public_key;
