+++ src/ecdsa/mod.rs
@@ -11,14 +11,12 @@ use elliptic_curve::{
     point::AffineCoordinates,
 };
 
-use frost_secp256k1::{
-    keys::SigningShare, Field, Secp256K1ScalarField, Secp256K1Sha256,
-};
+use frost_secp256k1::{keys::SigningShare, Field, Secp256K1ScalarField, Secp256K1Sha256};
 use k256::{AffinePoint, ProjectivePoint};
 
 use crate::crypto::ciphersuite::{BytesOrder, Ciphersuite, ScalarSerializationFormat};
 
-pub type KeygenOutput = crate::KeygenOutput::<Secp256K1Sha256>;
+pub type KeygenOutput = crate::KeygenOutput<Secp256K1Sha256>;
 
 pub type Scalar = <Secp256K1ScalarField as Field>::Scalar;
 pub type CoefficientCommitment = frost_core::keys::CoefficientCommitment<Secp256K1Sha256>;
