+++ src/eddsa/mod.rs
@@ -7,7 +7,7 @@ mod test;
 use crate::crypto::ciphersuite::{BytesOrder, Ciphersuite, ScalarSerializationFormat};
 use frost_ed25519::Ed25519Sha512;
 
-pub type KeygenOutput = crate::KeygenOutput::<Ed25519Sha512>;
+pub type KeygenOutput = crate::KeygenOutput<Ed25519Sha512>;
 
 impl ScalarSerializationFormat for Ed25519Sha512 {
     fn bytes_order() -> BytesOrder {
