+++ src/eddsa/test.rs
@@ -1,11 +1,9 @@
+use crate::crypto::hash::HashOutput;
 use crate::eddsa::sign::{sign, SignatureOutput};
 use crate::eddsa::KeygenOutput;
 use crate::participants::ParticipantList;
 use crate::protocol::{run_protocol, Participant, Protocol};
-use crate::crypto::hash::HashOutput;
-use crate::{keygen, reshare, refresh};
 
-use frost_ed25519::VerifyingKey;
 use rand_core::{OsRng, RngCore};
 use std::error::Error;
 
@@ -49,93 +47,6 @@ pub(crate) fn build_key_packages_with_dealer(
             )
         })
         .collect::<Vec<_>>()
-}
-
-/// runs distributed keygen
-pub(crate) fn run_keygen(
-    participants: &[Participant],
-    threshold: usize,
-) -> Result<Vec<(Participant, KeygenOutput)>, Box<dyn Error>> {
-    let mut protocols: Vec<(Participant, Box<dyn Protocol<Output = KeygenOutput>>)> =
-        Vec::with_capacity(participants.len());
-
-    for p in participants.iter() {
-        let protocol = keygen(participants, *p, threshold)?;
-        protocols.push((*p, Box::new(protocol)));
-    }
-
-    let result = run_protocol(protocols)?;
-    Ok(result)
-}
-
-/// runs distributed refresh
-pub(crate) fn run_refresh(
-    participants: &[Participant],
-    keys: Vec<(Participant, KeygenOutput)>,
-    threshold: usize,
-) -> Result<Vec<(Participant, KeygenOutput)>, Box<dyn Error>> {
-    let mut protocols: Vec<(Participant, Box<dyn Protocol<Output = KeygenOutput>>)> =
-        Vec::with_capacity(participants.len());
-
-    for (p, out) in keys.iter() {
-        let protocol = refresh(
-            Some(out.private_share),
-            out.public_key,
-            participants,
-            threshold,
-            *p,
-        )?;
-        protocols.push((*p, Box::new(protocol)));
-    }
-
-    let result = run_protocol(protocols)?;
-    Ok(result)
-}
-
-/// runs distributed reshare
-pub(crate) fn run_reshare(
-    participants: &[Participant],
-    pub_key: &VerifyingKey,
-    keys: Vec<(Participant, KeygenOutput)>,
-    old_threshold: usize,
-    new_threshold: usize,
-    new_participants: Vec<Participant>,
-) -> Result<Vec<(Participant, KeygenOutput)>, Box<dyn Error>> {
-    assert!(!new_participants.is_empty());
-    let mut setup: Vec<_> = vec![];
-
-    for new_participant in &new_participants {
-        let mut is_break = false;
-        for (p, k) in &keys {
-            if p == new_participant {
-                setup.push((*p, (Some(k.private_share), k.public_key)));
-                is_break = true;
-                break;
-            }
-        }
-        if !is_break {
-            setup.push((*new_participant, (None, *pub_key)));
-        }
-    }
-
-    let mut protocols: Vec<(Participant, Box<dyn Protocol<Output = KeygenOutput>>)> =
-        Vec::with_capacity(participants.len());
-
-    for (p, out) in setup.iter() {
-        let protocol = reshare(
-            participants,
-            old_threshold,
-            out.0,
-            out.1,
-            &new_participants,
-            new_threshold,
-            *p,
-        )?;
-        protocols.push((*p, Box::new(protocol)));
-    }
-
-    let result = run_protocol(protocols)?;
-    Ok(result)
 }
 
 pub(crate) fn test_run_signature_protocols(
@@ -185,20 +96,4 @@ pub(crate) fn test_run_signature_protocols(
     }
 
     Ok(run_protocol(protocols)?)
-}
-
-/// Assert that each participant has the same view of the public key
-pub(crate) fn assert_public_key_invariant(
-    participants: &[(Participant, KeygenOutput)],
-) -> Result<(), Box<dyn Error>> {
-    let public_key_package = participants.first().unwrap().1.public_key;
-
-    if participants
-        .iter()
-        .any(|(_, key_pair)| key_pair.public_key != public_key_package)
-    {
-        panic!("public key package is not the same for all participants");
-    }
-
-    Ok(())
 }
