+++ src/lib.rs
@@ -21,25 +21,24 @@ pub struct KeygenOutput<C: Ciphersuite> {
     pub public_key: VerifyingKey<C>,
 }
 
-
 /// Generic key generation function agnostic of the curve
-pub fn keygen<C:Ciphersuite>(
+pub fn keygen<C: Ciphersuite>(
     participants: &[Participant],
     me: Participant,
     threshold: usize,
 ) -> Result<impl Protocol<Output = KeygenOutput<C>>, InitializationError>
-where <<C::Group as frost_core::Group>::Field as frost_core::Field>::Scalar: Send,
-<C::Group as frost_core::Group>::Element: Send,
+where
+    <C::Group as frost_core::Group>::Element: Send,
+    <<C::Group as frost_core::Group>::Field as frost_core::Field>::Scalar: Send,
 {
     let comms = Comms::new();
     let participants = assert_keygen_invariants(participants, me, threshold)?;
-    let fut =
-      do_keygen::<C>(comms.shared_channel(), participants, me, threshold);
+    let fut = do_keygen::<C>(comms.shared_channel(), participants, me, threshold);
     Ok(make_protocol(comms, fut))
 }
 
 /// Performs the key reshare protocol
-pub fn reshare<C:Ciphersuite>(
+pub fn reshare<C: Ciphersuite>(
     old_participants: &[Participant],
     old_threshold: usize,
     old_signing_key: Option<SigningShare<C>>,
@@ -48,8 +47,9 @@ pub fn reshare<C:Ciphersuite>(
     new_threshold: usize,
     me: Participant,
 ) -> Result<impl Protocol<Output = KeygenOutput<C>>, InitializationError>
-where <<C::Group as frost_core::Group>::Field as frost_core::Field>::Scalar: Send,
-<C::Group as frost_core::Group>::Element: Send,
+where
+    <C::Group as frost_core::Group>::Element: Send,
+    <<C::Group as frost_core::Group>::Field as frost_core::Field>::Scalar: Send,
 {
     let comms = Comms::new();
     let threshold = new_threshold;
@@ -73,17 +73,17 @@ where <<C::Group as frost_core::Group>::Field as frost_core::Field>::Scalar: Sen
     Ok(make_protocol(comms, fut))
 }
 
-
 /// Performs the refresh protocol
-pub fn refresh<C:Ciphersuite>(
+pub fn refresh<C: Ciphersuite>(
     old_signing_key: Option<SigningShare<C>>,
     old_public_key: VerifyingKey<C>,
     new_participants: &[Participant],
     new_threshold: usize,
     me: Participant,
 ) -> Result<impl Protocol<Output = KeygenOutput<C>>, InitializationError>
-where <<C::Group as frost_core::Group>::Field as frost_core::Field>::Scalar: Send,
-<C::Group as frost_core::Group>::Element: Send,
+where
+    <C::Group as frost_core::Group>::Element: Send,
+    <<C::Group as frost_core::Group>::Field as frost_core::Field>::Scalar: Send,
 {
     if old_signing_key.is_none() {
         return Err(InitializationError::BadParameters(format!(
@@ -111,8 +111,6 @@ where <<C::Group as frost_core::Group>::Field as frost_core::Field>::Scalar: Sen
     );
     Ok(make_protocol(comms, fut))
 }
-
-
 
 // Libraries calls
 use crypto::ciphersuite::Ciphersuite;
