+++ src/test.rs
@@ -2,13 +2,12 @@
 // in the implemented schemes testing cases
 
 use rand_core::{OsRng, RngCore};
+use std::error::Error;
 
-use crate::KeygenOutput;
-use crate::protocol::Participant;
-use crate::Ciphersuite;
+use crate::protocol::{run_protocol, Participant, Protocol};
+use crate::{keygen, refresh, reshare, Ciphersuite, KeygenOutput, VerifyingKey};
 
 // +++++++++++++++++ Participants Utilities +++++++++++++++++ //
-
 /// Generates a vector of participants
 /// enumerated from 0 to number
 pub fn generate_participants(number: usize) -> Vec<Participant> {
@@ -25,4 +24,120 @@ pub fn generate_random_participants(number: usize) -> Vec<Participant> {
         .collect::<Vec<_>>();
     participants.sort();
     participants
+}
+
+// +++++++++++++++++ DKG Functions +++++++++++++++++ //
+/// Runs distributed keygen
+pub(crate) fn run_keygen<C: Ciphersuite>(
+    participants: &[Participant],
+    threshold: usize,
+) -> Result<Vec<(Participant, KeygenOutput<C>)>, Box<dyn Error>>
+where
+    <C::Group as frost_core::Group>::Element: Send,
+    <<C::Group as frost_core::Group>::Field as frost_core::Field>::Scalar: Send,
+{
+    let mut protocols: Vec<(Participant, Box<dyn Protocol<Output = KeygenOutput<C>>>)> =
+        Vec::with_capacity(participants.len());
+
+    for p in participants {
+        let protocol = keygen::<C>(participants, *p, threshold)?;
+        protocols.push((*p, Box::new(protocol)));
+    }
+
+    let result = run_protocol(protocols)?;
+    Ok(result)
+}
+
+/// Runs distributed refresh
+pub(crate) fn run_refresh<C: Ciphersuite>(
+    participants: &[Participant],
+    keys: Vec<(Participant, KeygenOutput<C>)>,
+    threshold: usize,
+) -> Result<Vec<(Participant, KeygenOutput<C>)>, Box<dyn Error>>
+where
+    <C::Group as frost_core::Group>::Element: Send,
+    <<C::Group as frost_core::Group>::Field as frost_core::Field>::Scalar: Send,
+{
+    let mut protocols: Vec<(Participant, Box<dyn Protocol<Output = KeygenOutput<C>>>)> =
+        Vec::with_capacity(participants.len());
+
+    for (p, out) in keys.iter() {
+        let protocol = refresh::<C>(
+            Some(out.private_share),
+            out.public_key,
+            participants,
+            threshold,
+            *p,
+        )?;
+        protocols.push((*p, Box::new(protocol)));
+    }
+
+    let result = run_protocol(protocols)?;
+    Ok(result)
+}
+
+/// runs distributed reshare
+pub(crate) fn run_reshare<C: Ciphersuite>(
+    participants: &[Participant],
+    pub_key: &VerifyingKey<C>,
+    keys: Vec<(Participant, KeygenOutput<C>)>,
+    old_threshold: usize,
+    new_threshold: usize,
+    new_participants: Vec<Participant>,
+) -> Result<Vec<(Participant, KeygenOutput<C>)>, Box<dyn Error>>
+where
+    <C::Group as frost_core::Group>::Element: Send,
+    <<C::Group as frost_core::Group>::Field as frost_core::Field>::Scalar: Send,
+{
+    assert!(!new_participants.is_empty());
+    let mut setup: Vec<_> = vec![];
+
+    for new_participant in &new_participants {
+        let mut is_break = false;
+        for (p, k) in &keys {
+            if p == new_participant {
+                setup.push((*p, (Some(k.private_share), k.public_key)));
+                is_break = true;
+                break;
+            }
+        }
+        if !is_break {
+            setup.push((*new_participant, (None, *pub_key)));
+        }
+    }
+
+    let mut protocols: Vec<(Participant, Box<dyn Protocol<Output = KeygenOutput<C>>>)> =
+        Vec::with_capacity(participants.len());
+
+    for (p, out) in setup.iter() {
+        let protocol = reshare(
+            participants,
+            old_threshold,
+            out.0,
+            out.1,
+            &new_participants,
+            new_threshold,
+            *p,
+        )?;
+        protocols.push((*p, Box::new(protocol)));
+    }
+
+    let result = run_protocol(protocols)?;
+    Ok(result)
+}
+
+/// Assert that each participant has the same view of the public key
+pub(crate) fn assert_public_key_invariant<C: Ciphersuite>(
+    participants: &[(Participant, KeygenOutput<C>)],
+) -> Result<(), Box<dyn Error>> {
+    let public_key_package = participants.first().unwrap().1.public_key;
+
+    if participants
+        .iter()
+        .any(|(_, key_pair)| key_pair.public_key != public_key_package)
+    {
+        panic!("public key package is not the same for all participants");
+    }
+
+    Ok(())
 }
